Synchronize requests on body using HTMX's drop strategy
Docd on sync startegies: https://htmx.org/attributes/hx-sync/
From the HTMX documentation on hx-sync I have gotten an erroneous notion that HTMX's "drop" strategy is by default set in the whole project.

But apparently "drop" is the default only if you use "hx-sync" attribute and don't provide any strategy, ie 'hx-sync' attribute must be used on the tag in order for HTMX to start synchronizing requests on (and within) this tag.

Synchronization is not a global config option in HTMX!