Shorten downtime info in the column
Uses a helper where implementation is copied from curitz.
Formatting downtime info in the column this way removes the stale-data-problem for this column completely. This is because there are no more redundant time units (that need constant updates) that are shown in the column, so whenever downtime for event is changed this info is updated in the specific row via NTIE.