Avoid exception if UpdateHandler is None on update
Try reconnecting UpdateHandler if None on update. Raise LostConnectionError on fail.

Also removes a try/except that was a vestige of full refresh. It is not relevant when using the UpdateHandler.

This is really a workaround, we do not yet know *why* the UpdateHandler can be None at this point in the flow.